refactor(fetch): split fetch.rs below 400 lines
Extract CDP/browser lifecycle into cdp.rs + cdp/launch.rs and HTTP download/decode into download.rs, then distribute the monolithic fetch/tests.rs across the four modules so each test sits beside its subject. All fetch files are now <=400 lines (fetch.rs 384, cdp.rs 273, launch.rs 247, download.rs 168).

Behavior unchanged: 449 tests (default) / 454 (js-rendering) pass, identical to the pre-split counts (bijective relocation, no test added or dropped).
